Mejora carga de imagenes update y create

+++ controllers/productoController.js
@@ -16,17 +16,25 @@ const createProductos = async (req, res) => {
         const { nombre, description, precio, stock } = req.body;
         let image;
         
-        if (req.files.image) {
-            const result = await uploadImage(req.files.image.tempFilePath);
-            await fs.remove(req.files.image.tempFilePath);
-
+        if(req.files!==null){
+            if (req.files.image) {
+                const result = await uploadImage(req.files.image.tempFilePath);
+                await fs.remove(req.files.image.tempFilePath);
+    
+                image = {
+                    url: result.secure_url,
+                    public_id: result.public_id,
+                };
+                
+                //console.log(result);
+            }
+        }else{
             image = {
-                url: result.secure_url,
-                public_id: result.public_id,
+                url: "https://res.cloudinary.com/dgi1aoxog/image/upload/v1668981098/productos/productoblacoynegro_ty6jw2.jpg",
+                public_id: "productos/productoblacoynegro_ty6jw2",
             };
-            
-            //console.log(result);
         }
+        
 
         const Newproducto = new Producto({ nombre, description, precio, image, stock });
         await Newproducto.save();
@@ -36,6 +44,20 @@ const createProductos = async (req, res) => {
     } catch (error) {
         console.log(error);
         return res.status(500).json({ msg: error.message });
+    }
+};
+
+const getProducto = async (req, res) => {
+    try {
+        const OneProduct = await Producto.findById(req.params.id);
+
+        if (!OneProduct) {
+            return res.sendStatus(404);
+        } else {
+            return res.json(OneProduct);
+        }
+    } catch (error) {
+        return res.status(500).json({ message: error.message });
     }
 };
 
@@ -54,29 +76,47 @@ const updateProductos = async (req, res) => {
 
     try {
         const updateProducto = await Producto.findById(id);
-        // console.log(updatePlato);
+        //console.log(updateProducto);
 
         updateProducto.nombre = nombre;
         updateProducto.description = description;
         updateProducto.precio = precio;
-        updateProducto.precio = stock;
+        updateProducto.stock = stock;
 
-        if (req.files.image) {
+        if(req.files!==null){
+            if (req.files.image) {
 
-            await deleteImage(updateProducto.image.public_id);
+                if (updateProducto.image.public_id && updateProducto.image.public_id !== "productos/productoblacoynegro_ty6jw2") {
+                    await deleteImage(updateProducto.image.public_id);
+                }
+    
+                const result = await uploadImage(req.files.image.tempFilePath);
 
-            const result = await uploadImage(req.files.image.tempFilePath);
-            await fs.remove(req.files.image.tempFilePath);
+                await fs.remove(req.files.image.tempFilePath);
+                
+                updateProducto.image = {
+                    url: result.secure_url,
+                    public_id: result.public_id,
+                }; 
+            }else{
+                updateProducto.image = {
+                    url: "https://res.cloudinary.com/dgi1aoxog/image/upload/v1668981098/productos/productoblacoynegro_ty6jw2.jpg",
+                    public_id: "productos/productoblacoynegro_ty6jw2",
+                };
+            }
+        }else{
+            if (!updateProducto.image.public_id){
+                updateProducto.image = {
+                    url: "https://res.cloudinary.com/dgi1aoxog/image/upload/v1668981098/productos/productoblacoynegro_ty6jw2.jpg",
+                    public_id: "productos/productoblacoynegro_ty6jw2",
+                };
+            }
             
-            updateProducto.image = {
-                url: result.secure_url,
-                public_id: result.public_id,
-            }; 
-
-            await updateProducto.save();
-
-            return res.status(204).json(updatePlato);
         }
+
+        await updateProducto.save();
+        return res.status(204).json(updateProducto);
+        
     } catch (error) {
         console.log(error.message);  
     }
@@ -91,31 +131,19 @@ const deleteProductos = async (req, res) => {
             return res.sendStatus(404);
         } else {
 
-            if (productRemoved.image.public_id) {
+            if (productRemoved.image.public_id && productRemoved.image.public_id !== "productos/productoblacoynegro_ty6jw2") {
                 await deleteImage(productRemoved.image.public_id);
             }
 
             // return res.sendStatus(204);
-            return res.status(500).json({ msg: "Producto Eliminado correctamente" });    
+            return res.status(200).json({ msg: "Producto Eliminado correctamente" });    
         }
     } catch (error) {
         return res.status(500).json({ message: error.message });
     }
 };
 
-const getProducto = async (req, res) => {
-    try {
-        const OneProduct = await Producto.findById(req.params.id);
 
-        if (!OneProduct) {
-            return res.sendStatus(404);
-        } else {
-            return res.json(OneProduct);
-        }
-    } catch (error) {
-        return res.status(500).json({ message: error.message });
-    }
-};
 
 export {
     prueba,
